chore(storage): restructure source files into web/cookies/ and src/
- web/cookie-attributes.mts  → web/cookies/cookie-helper.mts
- web/cookie-serialize.mts   → serializer.mts
- web/cookie-storage.mts     → web/cookies/cookie-storage.mts
